pyproject.toml
[tool.poetry]
name = "staba"
version = "0.1.0"
description = ""
authors = ["Fatumayattani <fyattani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uagents = "^0.22.3"
python-dotenv = "^1.1.0"
google-generativeai = "^0.8.4"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
